Разделил main.js  на модули
